\name{plot_diff_sumscore}
\alias{plot_diff_sumscore}
\title{Plot Differences in Sum of Scores for Ising Model Simulations}
\description{
The \code{plot_diff_sumscore} function creates a box plot or violin plot displaying the distribution of sum of scores for each intervention scenario in the Ising model simulations generated using the \code{snp_Ising} function. It also performs pairwise comparisons and displays the significant differences between the interventions. This function differs from the \code{snp_plot} function, which creates a line plot of the mean sum of scores with error bars.
}
\usage{
plot_diff_sumscore(simulations,
                  type = "parametric",
                  pairwise.display = "significant",
                  plot.type = "box",
                  showdots = FALSE,
                  ylab = "Sum score",
                  theme = ggthemes::theme_solarized(),
                  ...)
}
\arguments{
\item{simulations}{An object of class "multi easy ising simulations obj" generated by the \code{snp_Ising} function.}
\item{type}{A character value specifying the type of test to use for pairwise comparisons. Either "parametric" or "nonparametric". Default is "parametric".}
\item{pairwise.display}{A character value specifying how to display the pairwise comparison results. Either "all", "significant", or "none". Default is "significant".}
\item{plot.type}{A character value specifying the type of plot to create. Either "box" or "violin". Default is "box".}
\item{showdots}{A logical value indicating whether to show individual data points on the plot. Default is FALSE.}
\item{ylab}{A character value specifying the y-axis label. Default is "Sum score".}
\item{theme}{A ggplot2 theme for customizing the plot appearance. Default is \code{ggthemes::theme_solarized()}.}
\item{...}{Additional arguments to be passed to the \code{ggbetweenstats} function.}
}
\details{
The \code{plot_diff_sumscore} function takes the output from the \code{snp_Ising} function and creates a box plot or violin plot displaying the distribution of sum of scores for each intervention scenario. It also performs pairwise comparisons and displays the significant differences between the interventions.

The function uses the \code{ggstatsplot} package for creating the plot and allows customization of the plot appearance using the \code{theme} parameter.
}
\value{
A ggplot object representing the box or violin plot of the sum of scores for each intervention scenario, with pairwise comparison results displayed.
}
\seealso{
\code{\link{snp_Ising}}, \code{\link{ggplot2}}, \code{\link{ggthemes}}, \code{\link{snp_plot}}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{

}
\note{
%% further notes
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
\examples{
 plot.net.diff(simulations)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
