% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_bigcorPar.r
\name{snp_bigcorPar}
\alias{snp_bigcorPar}
\title{Title}
\usage{
snp_bigcorPar(x, size = 1000, verbose = TRUE, ncore = "all", ...)
}
\arguments{
\item{x}{The matrix for processing. Notably, the pairwise distance is calculated among column.}

\item{size}{Number of columns in each block. Default value is 1000.}

\item{ncore}{Number of CPU to use}

\item{...}{Other parameters for calculate distance, pass to cor()}
}
\value{
A distance matrix among columns
}
\description{
Title
}
