% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_dendrite.r
\name{snp_calculate_dendrite_feature}
\alias{snp_calculate_dendrite_feature}
\title{Calculate morphological features of dendrites}
\usage{
snp_calculate_dendrite_feature(neuronlist)
}
\arguments{
\item{neuronlist}{list of dendrites}
}
\value{
feature matrix
}
\description{
Calculate morphological features of dendrites
}
