\name{snp_graph}
\alias{snp_graph}
\title{Visualize Ising Model Simulations}
\description{
The \code{snp_graph} function creates a series of network graphs for the results of Ising model simulations generated using the \code{snp_Ising} function. The function provides a visualization of the relationships between variables in each intervention scenario, with optional community detection and various customization options.
}
\usage{
snp_graph(simulations,
          network_community_detection = FALSE,
          layout = "spring",
          fade = TRUE,
          negDashed = TRUE,
          edge.labels = FALSE,
          legend = FALSE,
          grah_nb_col = 3,
          ...)
}
\arguments{
\item{simulations}{An object of class "multi easy ising simulations obj" generated by the \code{snp_Ising} function.}
\item{network_community_detection}{A logical value. If \code{TRUE}, the function will perform community detection on the network graphs. Default is \code{FALSE}.}
\item{layout}{A character string specifying the layout algorithm for the network graphs. Default is "spring".}
\item{fade}{A logical value. If \code{TRUE}, the strength of the edges in the network graphs will be faded based on the edge weights. Default is \code{TRUE}.}
\item{negDashed}{A logical value. If \code{TRUE}, negative edges in the network graphs will be displayed as dashed lines. Default is \code{TRUE}.}
\item{edge.labels}{A logical value. If \code{TRUE}, edge labels will be displayed in the network graphs. Default is \code{FALSE}.}
\item{legend}{A logical value. If \code{TRUE}, a legend will be displayed for the network graphs. Default is \code{FALSE}.}
\item{grah_nb_col}{An integer value specifying the number of columns for arranging the network graphs. Default is 3.}
\item{...}{Additional arguments to be passed to the \code{qgraph} function.}
}
\details{
The \code{snp_graph} function takes the output from the \code{snp_Ising} function and creates a series of network graphs for each intervention scenario. The function uses the \code{qgraph} package for visualization and offers various customization options, such as layout algorithm, edge fading, and community detection.

The function returns a list containing the network graphs and, if community detection is enabled, a list of community membership information.
}
\value{
A list containing the following elements:
\describe{
\item{list_graph}{A list of network graphs for each intervention scenario.}
\item{list_NetworkCom (optional)}{A list of community membership information for each intervention scenario if \code{network_community_detection} is set to \code{TRUE}.}
}
}
\seealso{
\code{\link{snp_Ising}}, \code{\link{qgraph}}, \code{\link{igraph}}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%% ~put your name and email here ~
}
\note{
%% further notes
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
