\name{snp_Ising}
\alias{snp_Ising}
\title{Perform Multiple Ising Model Simulations}
\description{
The \code{snp_Ising} function allows users to perform multiple Ising model simulations by utilizing the \code{easy_ising_simulations} function. This function is designed to handle multiple input scenarios and run corresponding simulations in parallel.
}
\usage{
snp_Ising(df_list,
          nodes_to_influence_list,
          centrality_indices_list = NULL,
          relation_list = NULL,
          Fit_Ising_list = NULL,
          Fit_Ising_thresholds_list,
          n,
          thresholds_IsingSampler_list,
          beta_list)
}
\arguments{
\item{df_list}{A list of data frames containing the input data for the Ising model.}
\item{nodes_to_influence_list}{A list of numeric vectors or character vectors specifying the nodes to influence based on their centrality indices for each data frame in df_list.}
\item{centrality_indices_list}{A list of character strings specifying the centrality index to use for selecting nodes to influence for each data frame in df_list. Acceptable values include "Betweenness", "Closeness", "Strength", and "ExpectedInfluence". Default is NULL.}
\item{relation_list}{A list of character strings specifying the relation to use when selecting nodes based on their centrality indices for each data frame in df_list. Acceptable values include ">=" or "<=". Default is NULL.}
\item{Fit_Ising_list}{A list of objects created by the \code{IsingFit} function for each data frame in df_list. If NULL, the function will create new IsingFit objects based on the input data frames. Default is NULL.}
\item{Fit_Ising_thresholds_list}{A list of logical values indicating whether to use the original thresholds from the Fit_Ising objects for each data frame in df_list.}
\item{n}{A numeric value specifying the number of samples to simulate. Must be greater than 1.}
\item{thresholds_IsingSampler_list}{A list of numeric value indicating the extent of modification applied to the threshold level. This value can be a positive or negative integer, representing the number of standard deviations added or subtracted from the original threshold. For instance, a value of 1 signifies the addition of one standard deviation to the existing threshold, whereas a value of -0.5 denotes the removal of half a standard deviation from the original threshold.}
\item{beta_list}{A list of single numeric values representing the inverse temperature parameter for the Ising model for each data frame in df_list.}
}
\details{
The \code{snp_Ising} function is designed to perform multiple Ising model simulations using the \code{easy_ising_simulations} function. This function takes multiple input scenarios, such as different data frames, nodes to influence, centrality indices, and other parameters, and runs the corresponding simulations in parallel.

The function returns an object of class "multi easy ising simulations obj" containing a list of simulation results for each input scenario.
}
\value{
A list of class "multi easy ising simulations obj" with the following components:
\describe{
\item{simulations}{A list of simulation results generated by the \code{easy_ising_simulations} function for each input scenario.}
}
}
\seealso{
\code{\link{easy_ising_simulations}}, \code{\link{IsingFit}}, \code{\link{IsingSampler}}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Renaud Mabire-Yon (he/him/his) renaud.mabire.yon@gmail.com
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
\examples{
\dontrun{

## Load necessary libraries ##
library('IsingFit')
library('IsingSampler')
library('qgraph')
library('dplyr')

## Simulate dataset ##
## Input:
N <- 10 # Number of nodes
nSample <- 1000 # Number of samples

## Ising parameters:
Graph <- matrix(sample(0:1, N^2, TRUE, prob = c(0.8, 0.2)), N, N) * runif(N^2, 0.5, 2)
Graph <- pmax(Graph, t(Graph))
diag(Graph) <- 0
Thresh <- -rowSums(Graph) / 2

## Simulate:
Data <- IsingSampler(nSample, Graph, Thresh) %>% as.data.frame()

# The snp_Ising function is applied to this dataset four times with varying configurations.
# In each case, we focus on influencing nodes "V1" and "V2". The function is run with the same
# dataset, but with different threshold levels specified in the thresholds_IsingSampler_list
# argument, which contains values of 0.2, 0.5, 0.8, and 1.1. These values correspond to the
# adjustments made to the original threshold for each iteration of the function.
# By running the snp_Ising function with these specific settings, we aim to explore the effects
# of altering the threshold levels for the selected nodes. The results of the simulations will
# provide insights into how the model's behavior changes with varying threshold values, allowing
# us to better understand the Ising model's dynamics and potential real-world applications.


simulations <- snp_Ising(df_list = list(Data, Data, Data, Data),
                     nodes_to_influence_list = list(c("V8", "V10"),
                                                    c("V8", "V10"),
                                                    c("V8", "V10"),
                                                    c("V8", "V10")),
                     centrality_indices_list = NULL,
                     relation_list = NULL,
                     Fit_Ising_list = NULL,
                     Fit_Ising_thresholds_list = list(FALSE, FALSE, FALSE, FALSE),
                     n = rep(1000, 4),
                     thresholds_IsingSampler_list = list(0.1, 0.4, 0.7, 1),
                     beta_list = rep(1, 4))

# Summary of simulations
summary_snp(simulations)





}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
