\name{snp_plot}
\alias{snp_plot}
\title{Plot Sum of Scores for Ising Model Simulations}
\description{
The \code{snp_plot} function creates a line plot of the mean sum of scores for each intervention scenario in the Ising model simulations generated using the \code{snp_Ising} function. The plot includes error bars representing the confidence intervals, and displays ANOVA results in the subtitle.
}
\usage{
snp_plot(simulations, CI = 0.95, theme = ggthemes::theme_solarized())
}
\arguments{
\item{simulations}{An object of class "multi easy ising simulations obj" generated by the \code{snp_Ising} function.}
\item{CI}{A numeric value representing the desired confidence interval level. Default is 0.95.}
\item{theme}{A ggplot2 theme for customizing the plot appearance. Default is \code{ggthemes::theme_solarized()}.}
}
\details{
The \code{snp_plot} function takes the output from the \code{snp_Ising} function and creates a line plot of the mean sum of scores for each intervention scenario. The plot includes error bars representing the confidence intervals, and ANOVA results are displayed in the subtitle.

The function uses the \code{ggplot2} package for creating the plot and allows customization of the plot appearance using the \code{theme} parameter.
}
\value{
A ggplot object representing the line plot of the mean sum of scores for each intervention scenario, including error bars and ANOVA results.
}
\seealso{
\code{\link{snp_Ising}}, \code{\link{ggplot2}}, \code{\link{ggthemes}}, \code{\link{summary_snp}}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%% ~put your name and email here ~
}
\note{
%% further notes
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
