% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{standard_normal_plot}
\alias{standard_normal_plot}
\title{Descriptive statistics}
\usage{
standard_normal_plot(size = 11)
}
\arguments{
\item{size}{text size}
}
\description{
uses psych
}
\details{
returns xlsx
}
\examples{
standard_normal_plot()
}
