\name{summary_snp}
\alias{summary_snp}
\title{Summarize Results of Ising Model Simulations}
\description{
The \code{summary_snp} function computes summary statistics for the results of Ising model simulations generated using the \code{snp_Ising} function. This function calculates the mean, standard deviation, standard error, and confidence intervals for the sum scores of the simulated interventions.
}
\usage{
summary_snp(simulations, CI = 0.95)
}
\arguments{
\item{simulations}{An object of class "multi easy ising simulations obj" generated by the \code{snp_Ising} function.}
\item{CI}{A numeric value specifying the confidence level to compute the confidence interval. Default is 0.95, corresponding to a 95% confidence interval.}
}
\details{
The \code{summary_snp} function takes the output from the \code{snp_Ising} function and calculates summary statistics for each intervention scenario. It provides a concise summary of the sum scores of the interventions, making it easier to compare and analyze the results.

The function returns a data frame containing the intervention, number of samples, mean sum score, standard deviation, standard error, and the lower and upper bounds of the confidence interval.
}
\value{
A data frame with the following columns:
\describe{
\item{intervention}{The intervention scenario identifier.}
\item{n}{The number of samples in the intervention scenario.}
\item{sumscore_mean}{The mean sum score for the intervention scenario.}
\item{sd}{The standard deviation of the sum scores for the intervention scenario.}
\item{se}{The standard error of the mean sum score for the intervention scenario.}
\item{ciLower}{The lower bound of the confidence interval for the mean sum score.}
\item{ciUpper}{The upper bound of the confidence interval for the mean sum score.}
}
}
\seealso{
\code{\link{snp_Ising}}, \code{\link{Rmisc::summarySE}}, \code{\link{dplyr}}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%% ~put your name and email here ~
}
\note{
%% further notes
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
\examples{
\dontrun{

Load necessary libraries
library('IsingFit')
library('IsingSampler')
library('qgraph')
library('dplyr')
library('Rmisc')

Simulate dataset and run snp_Ising function
%% (Please refer to the snp_Ising documentation for details on this part) %%

Assume the 'results' object is generated from snp_Ising function
summary_results <- summary_snp(results)

Print the summary of simulations
print(summary_results)

}
}



