% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shared_allen_anno}
\alias{shared_allen_anno}
\title{shared_allen_anno}
\format{
A data frame with 1,327 rows and 5 columns:
\describe{
\item{CurrentID}
\item{ParentID}
}
}
\usage{
shared_allen_anno
}
\description{
allen ccf v3
}
\keyword{datasets}
