% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_computation.r
\name{snp_calculateProjections}
\alias{snp_calculateProjections}
\title{Calculate the axon lengths in annotated brain regions}
\usage{
snp_calculateProjections(
  PathOfNeuron,
  Hemisphere,
  StructureIDOfTargets,
  Annotation,
  Resolution,
  NumOfCores,
  OptionMetric = "LengthOfAxon"
)
}
\arguments{
\item{PathOfNeuron}{Full path of SWC file for each neuron}

\item{Hemisphere}{Which hemisphere does the neuron reside in. Note that neurons in right hemisphere were flipped to the left side.}

\item{StructureIDOfTargets}{Structure ID of brain regions or domains in the annotation.
Column name of returned matrix is the same as StuctureIDOfTargets.}

\item{Annotation}{3D array with each element assigned a positive value}

\item{Resolution}{Voxel size of each annotation, e.g., 100, 50, 25, 10 um}

\item{NumOfCores}{Number of cores for parallel computation}

\item{OptionMetric}{Default value is "LengthOfAxon". ("LengthOfAxon", "NumberOfEndNodes", "NumberOfBranchNodes")}
}
\value{
projection matrix
}
\description{
Calculate the axon lengths in annotated brain regions
}
\details{
Annotated brain regions can be any 3D array with
an unique index denoting a brain region or domain
}
