% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_util.r
\name{snp_getAcronymFromStructureID}
\alias{snp_getAcronymFromStructureID}
\title{Convert StructureID to Acronym}
\usage{
snp_getAcronymFromStructureID(structureList)
}
\arguments{
\item{structureList}{}
}
\description{
Convert StructureID to Acronym
}
