% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_util.r
\name{snp_getStructureIDFromAcronym}
\alias{snp_getStructureIDFromAcronym}
\title{convert Acronym to StructureID}
\usage{
snp_getStructureIDFromAcronym(acronymList, flag_debug = FALSE)
}
\arguments{
\item{acronymList}{}
}
\description{
This function returns the structureID of specified regions (acronym)
}
\examples{
snp::getStructureFromAcronym(c("l_Isocortex", "r_CP"))
}
