% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{standard_normal_plot}
\alias{standard_normal_plot}
\title{Plot standard normal plot}
\usage{
standard_normal_plot(size = 10)
}
\arguments{
\item{size}{text size}
}
\description{
Plot standard normal plot
}
\examples{
standard_normal_plot()
standard_normal_plot(size=5)
}
